# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: EZ/PZ API (the parts the native apps call)
  version: 0.1.0
  description: |
    The contract between the Flask API and the web and native clients. It holds
    the routes the apps call, one at a time as each arrives; a route missing from
    here is not yet something an app may rely on. `backend/tests/test_openapi_contract.py`
    fails when a response stops matching, so a server change that would break an
    installed app shows up as a test failure first.

    Store builds stay in the field for months. Changes here are additive: a new
    optional field, a new route. Removing or retyping a field breaks apps that
    cannot be updated, and needs `configVersion` bumped and `minAppVersion` raised.

    Every route is under `/api`. The base URL is the deployment's (the same host
    the web app uses); it is not part of the spec.
servers:
  - url: https://prod-ezpz-api.mcfadd.in
    description: Self-hosted production (Coolify behind Cloudflare Tunnel)
  - url: http://127.0.0.1:5000
    description: Local development
paths:
...
  /api/teams:
    get:
      operationId: listTeams
      summary: The caller's teams
      security:
        - bearer: []
      responses:
        "200":
          description: Teams, by name.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [teams]
                properties:
                  teams:
                    type: array
                    items: { $ref: "#/components/schemas/TeamSummary" }
    post:
      operationId: createTeam
      summary: Make a team; the caller owns it
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
      responses:
        "201":
          description: Made.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "400":
          $ref: "#/components/responses/PackError"
  /api/teams/{id}:
    parameters:
      - $ref: "#/components/parameters/TeamId"
    get:
      operationId: getTeam
      summary: A team and its members (members only)
      security:
        - bearer: []
      responses:
        "200":
          description: The team.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "404":
          $ref: "#/components/responses/PackError"
    put:
      operationId: renameTeam
      summary: Rename a team (owner or admin)
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
      responses:
        "200":
          description: The team.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "400":
          $ref: "#/components/responses/PackError"
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
    delete:
      operationId: deleteTeam
      summary: Delete a team (owner). Packs shared with it stop being shared, each saying so in its log
      security:
        - bearer: []
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
  /api/teams/{id}/members/{user_id}:
    parameters:
      - $ref: "#/components/parameters/TeamId"
      - $ref: "#/components/parameters/UserId"
    put:
      operationId: changeTeamMember
      summary: Change a member's role (owner). `owner` hands the team over; the old owner becomes an admin
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [owner, admin, member] }
      responses:
        "200":
          description: The team.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Team" }
        "400":
          $ref: "#/components/responses/PackError"
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
        "409":
          $ref: "#/components/responses/PackError"
    delete:
      operationId: removeTeamMember
      summary: Leave (yourself), or remove someone (owner; an admin removes members only)
      security:
        - bearer: []
      responses:
        "200":
          description: Removed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Status" }
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
        "409":
          $ref: "#/components/responses/PackError"
  /api/teams/{id}/invites:
    parameters:
      - $ref: "#/components/parameters/TeamId"
    get:
      operationId: listTeamInvites
      summary: Invitations not yet answered (owner or admin)
      security:
        - bearer: []
      responses:
        "200":
          description: Invitations.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteList" }
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
    post:
      operationId: inviteToTeam
      summary: Invite by email, or with no email make a single-use link (owner or admin)
      description: A link's `token` is in this response only; it is stored hashed and never shown again.
      security:
        - bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                role: { type: string, enum: [member, admin], default: member, description: Only the owner gives `admin`. }
      responses:
        "201":
          description: Invited.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteReply" }
        "400":
          $ref: "#/components/responses/PackError"
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
        "409":
          $ref: "#/components/responses/PackError"
        "429":
          $ref: "#/components/responses/PackError"
  /api/teams/{id}/invites/{invite_id}:
    parameters:
      - $ref: "#/components/parameters/TeamId"
      - $ref: "#/components/parameters/InviteId"
    delete:
      operationId: revokeTeamInvite
      summary: Withdraw an invitation (owner or admin)
      security:
        - bearer: []
      responses:
        "200":
          description: Withdrawn.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteOnly" }
        "403":
          $ref: "#/components/responses/PackError"
        "404":
          $ref: "#/components/responses/PackError"
  /api/users/search:
    get:
      operationId: searchUsers
      summary: People who share a team with the caller, by name or sign-in email (20 at most)
      description: There is no open directory. A `.mil` address someone did not sign in with is never searched or shown.
      security:
        - bearer: []
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 100 }
      responses:
        "200":
          description: Matches, by name.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [users]
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/UserBrief" }
        "400":
          $ref: "#/components/responses/PackError"
  /api/invites:
    get:
      operationId: myInvites
      summary: Invitations waiting for the caller, to their sign-in address or their verified `.mil` one
      security:
        - bearer: []
      responses:
        "200":
          description: Pending, unexpired invitations, newest first.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteList" }
  /api/invites/{invite_id}/accept:
    parameters:
      - $ref: "#/components/parameters/InviteId"
    post:
      operationId: acceptInvite
      summary: Accept an invitation addressed to the caller
      security:
        - bearer: []
      responses:
        "200":
          description: Joined.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteAccepted" }
        "404":
          $ref: "#/components/responses/PackError"
        "410":
          $ref: "#/components/responses/PackError"
  /api/invites/{invite_id}/decline:
    parameters:
      - $ref: "#/components/parameters/InviteId"
    post:
      operationId: declineInvite
      summary: Decline an invitation addressed to the caller
      security:
        - bearer: []
      responses:
        "200":
          description: Declined.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteOnly" }
        "404":
          $ref: "#/components/responses/PackError"
        "410":
          $ref: "#/components/responses/PackError"
  /api/invites/accept:
    post:
      operationId: acceptInviteByToken
      summary: Accept from an emailed or shared link, whatever address the caller signed in with
      security:
        - bearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: "The `invite` query parameter of the link." }
      responses:
        "200":
          description: Joined.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InviteAccepted" }
        "404":
          $ref: "#/components/responses/PackError"
        "410":
          $ref: "#/components/responses/PackError"
components:
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema: { type: integer }
    TeamId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    InviteId:
      name: invite_id
      in: path
      required: true
      schema: { type: integer }
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: The `access_token` from sign-in or refresh.
...
  responses:
...
    PackError:
      description: An error from the pack and team routes. `code` is stable and meant for the app.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PackError"
  schemas:
...
    Person:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id: { type: integer, nullable: true, description: "Null once the account is deleted (the name is kept as it was)." }
        name: { type: string }
    Status:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: { type: string, enum: [deleted, removed] }
    PackError:
      type: object
      required: [error]
      description: |
        `error` is for people; `code` is stable and meant for the app: `pack_not_found`,
        `pack_read_only` (a viewer), `owner_only`, `pack_finished` (423, with who finished it and
        when), `invalid_batch`, `invalid_op` (with `index` and the `reason` from
        contracts/fixtures/packs/ops.json, or `bad_client_op_id`, `bad_summary`), `item_too_large`
        (413, with `item`), `invalid_name`, `invalid_description`, `invalid_uuid`, `uuid_taken`,
        `not_in_team`, `invalid_role`, `invalid_source`, `invalid_item`, `source_not_found`,
        `mission_not_supported`, `unreadable_source`, `item_exists`, `item_not_found`,
        `not_your_original`, `original_gone`, `empty_point_set`, `feature_disabled`,
        `not_a_teammate`, `already_member`, `member_not_found`, `owner_must_transfer`,
        `invalid_email`, `invite_not_found`, `invite_gone` / `invite_expired` (410),
        `rate_limited` (429), `team_not_found`, `team_managers_only`, `invalid_query`.
      properties:
        error: { type: string }
        code: { type: string }
        index: { type: integer }
        reason: { type: string }
        item: { type: string }
        finished_at: { type: string, nullable: true }
        finished_by:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Person" }]
    PackTeam:
      type: object
      additionalProperties: false
      required: [id, name, member_count, role]
      properties:
        id: { type: integer }
        name: { type: string }
        member_count: { type: integer, minimum: 0 }
        role: { type: string, enum: [editor, viewer], description: What the team's members may do in the pack. }
    PackSummary:
      type: object
      additionalProperties: false
      required: [uuid, name, description, status, role, owner, team, head_seq, member_count, item_count,
                 finished_at, finished_by, created_at, updated_at]
      properties:
        uuid: { type: string }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, finished], description: "finished: read-only for everyone." }
        role: { type: string, enum: [owner, editor, viewer], description: The caller's. }
        owner: { $ref: "#/components/schemas/Person" }
        team:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/PackTeam" }]
        head_seq: { type: integer, minimum: 0, description: The number of the pack's latest event. }
        member_count: { type: integer, minimum: 0, description: "Members in their own right, not counting a shared team." }
        item_count: { type: integer, minimum: 0 }
        finished_at: { type: string, nullable: true }
        finished_by:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Person" }]
        created_at: { type: string }
        updated_at: { type: string }
...
    Invite:
      type: object
      additionalProperties: false
      required: [id, email, role, status, invited_by, expires_at, created_at, pack, team]
      properties:
        id: { type: integer }
        email: { type: string, nullable: true, description: "Null for a team's single-use link." }
        role: { type: string, enum: [editor, viewer, admin, member] }
        status: { type: string, enum: [pending, accepted, declined, revoked, expired] }
        invited_by:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Person" }]
        expires_at: { type: string }
        created_at: { type: string }
        pack:
          type: object
          nullable: true
          additionalProperties: false
          required: [uuid, name]
          properties:
            uuid: { type: string }
            name: { type: string }
        team:
          type: object
          nullable: true
          additionalProperties: false
          required: [id, name]
          properties:
            id: { type: integer }
            name: { type: string }
    InviteOnly:
      type: object
      additionalProperties: false
      required: [invite]
      properties:
        invite: { $ref: "#/components/schemas/Invite" }
    InviteList:
      type: object
      additionalProperties: false
      required: [invites]
      properties:
        invites:
          type: array
          items: { $ref: "#/components/schemas/Invite" }
    InviteReply:
      type: object
      additionalProperties: false
      required: [invite]
      properties:
        invite: { $ref: "#/components/schemas/Invite" }
        email_sent: { type: boolean, description: Present for an emailed invitation. }
        token: { type: string, description: "A team link's token: shown here once, never again." }
    InviteAccepted:
      type: object
      additionalProperties: false
      required: [invite, pack, team]
      properties:
        invite: { $ref: "#/components/schemas/Invite" }
        pack:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/PackSummary" }]
        team:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/TeamSummary" }]
    TeamSummary:
      type: object
      additionalProperties: false
      required: [id, name, role, member_count, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string, enum: [owner, admin, member], description: The caller's. }
        member_count: { type: integer, minimum: 1 }
        created_at: { type: string }
    Team:
      type: object
      additionalProperties: false
      required: [id, name, role, member_count, created_at, members]
      properties:
        id: { type: integer }
        name: { type: string }
        role: { type: string, enum: [owner, admin, member] }
        member_count: { type: integer, minimum: 1 }
        created_at: { type: string }
        members:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [user_id, name, email, role, joined_at]
            properties:
              user_id: { type: integer }
              name: { type: string }
              email: { type: string }
              role: { type: string, enum: [owner, admin, member] }
              joined_at: { type: string }
    UserBrief:
      type: object
      additionalProperties: false
      required: [id, name, email]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, description: The sign-in address. }